feat: merge k8s-launcher with k8s-helper
1. merge k8s-launcher with k8s-helper
2. move kubernetes package import to where it will be used

3. hack/install-k8s-tools.sh -> ding/scripts/install-k8s-tools.sh
