+++ .github/workflows/unit_test.yml
@@ -23,7 +23,7 @@ jobs:
         run: |
           python -m pip install .
           python -m pip install ".[test,k8s]"
-          ./hack/install-k8s-tools.sh
+          ./ding/scripts/install-k8s-tools.sh
           make unittest
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
@@ -51,5 +51,5 @@ jobs:
         run: |
           python -m pip install .
           python -m pip install ".[test,k8s]"
-          ./hack/install-k8s-tools.sh
+          ./ding/scripts/install-k8s-tools.sh
           make benchmark
