+++ ding/utils/__init__.py
@@ -9,8 +9,8 @@ from .file_helper import read_file, save_file, remove_file
 from .import_helper import try_import_ceph, try_import_mc, try_import_link, import_module, try_import_redis, \
     try_import_rediscluster
 from .k8s_helper import get_operator_server_kwargs, exist_operator_server, DEFAULT_K8S_COLLECTOR_PORT, \
-    DEFAULT_K8S_LEARNER_PORT, DEFAULT_K8S_AGGREGATOR_SLAVE_PORT, DEFAULT_K8S_COORDINATOR_PORT, pod_exec_command
-from .k8s_launcher import K8sLauncher
+    DEFAULT_K8S_LEARNER_PORT, DEFAULT_K8S_AGGREGATOR_SLAVE_PORT, DEFAULT_K8S_COORDINATOR_PORT, pod_exec_command, \
+    K8sLauncher
 from .orchestrator_launcher import OrchestratorLauncher
 from .lock_helper import LockContext, LockContextType, get_file_lock, get_rw_file_lock
 from .log_helper import build_logger, DistributionTimeImage, pretty_print, LoggerFactory
